# Performance: Fixes Reduces AuthorizationHelper memory allocations