--- a/Microsoft.Azure.Cosmos/src/SecureStringHMACSHA256Helper.cs
+++ b/Microsoft.Azure.Cosmos/src/SecureStringHMACSHA256Helper.cs
@@ -75,7 +75,7 @@ namespace Microsoft.Azure.Cosmos
         /// SHA256 HMAC of the payload, and destroy the native memory containing the decoded key.
         /// </summary>
         /// <param name="bytesToHash">payload that is hashed</param>
-        public byte[] ComputeHash(MemoryStream bytesToHash)
+        public byte[] ComputeHash(ArraySegment<byte> bytesToHash)
         {
             IntPtr hashHandle = IntPtr.Zero;
 
@@ -95,15 +95,15 @@ namespace Microsoft.Azure.Cosmos
             }
         }
 
-        private void AddData(IntPtr hashHandle, MemoryStream dataStream)
+        private void AddData(IntPtr hashHandle, ArraySegment<byte> dataStream)
         {
-            byte[] data = dataStream.GetBuffer();
+            byte[] data = dataStream.Array;
             GCHandle h = GCHandle.Alloc(data, GCHandleType.Pinned);
             try
             {
                 int status = NativeMethods.BCryptHashData(hashHandle,
                     h.AddrOfPinnedObject(),
-                    (uint)data.Length,
+                    (uint)dataStream.Count,
                     0);
                 if (status != 0)
                 {