--- a/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Benchmarks/MasterKeyAuthorizationBenchmark.cs
+++ b/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Benchmarks/MasterKeyAuthorizationBenchmark.cs
@@ -45,7 +45,7 @@ namespace Microsoft.Azure.Cosmos.Performance.Tests
             AuthorizationHelper.GenerateKeyAuthorizationSignature(verb, resourceId, resourceType, this.testHeaders, this.authKeyHashFunction, out payload);
 
 #pragma warning disable CS0642 // Possible mistaken empty statement
-            //using (payload) ;
+            using (payload) ;
 #pragma warning restore CS0642 // Possible mistaken empty statement
         }
     }