Including exclusive auth performance micro benchmark.
Changes includes: 

1. Caching the HMACSHA256
2. Contract adjusted to avoid leaking bytes out

After the changes
RPS: 2-3% UP
Latency: 3-6% UP
GEN0: 30% DOWN
Allocations: 28% DOWN

|                    Method |     Mean |     Error |    StdDev |       Q3 |      P80 |      P85 |      P90 |      P95 |     P100 |      Op/s |  Gen 0 | Allocated |
|-------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|---------:|----------:|-------:|----------:|
| CreateSignatureGeneration | 3.579 us | 0.0336 us | 0.0280 us | 3.603 us | 3.605 us | 3.612 us | 3.612 us | 3.614 us | 3.618 us | 279,387.0 | 0.0458 |     896 B |
|   ReadSignatureGeneration | 3.542 us | 0.0490 us | 0.0458 us | 3.576 us | 3.577 us | 3.580 us | 3.598 us | 3.617 us | 3.633 us | 282,301.5 | 0.0458 |     904 B |

Before changes:
Branch with-out Auth changes & perf benchmark: **users/kirankk/auth_baseline**
**NOTE**: Please change the target to netcoreapp3.1 

| OS | Allocations %Change | GC %Change | RPs %Change |
|---------------- |---------:|----------:|----------:|
| WINDOWS | ~174B Less | 11% Less GC | 7-9% less|
| LINUX | ~174B Less | 10% Less GC | 13-15% less | 

Also the latency is higher with the new optimization.

- [ ] Is high latency and less RPS results of high CPU work? 
- [ ] If so if it worth trading for allocations?

>   Op/s/Rps           : Operation per second
>   Gen 0                : GC Generation 0 collects per 1000 operations
>   Gen 1                : GC Generation 1 collects per 1000 operations
>   Gen 2                : GC Generation 2 collects per 1000 operations
>   Allocated            : Allocated memory per single operation (managed only, inclusive, 1KB = 1024B)

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.302
  [Host]     : .NET Core 3.1.6 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.31603), X64 RyuJIT
  DefaultJob : .NET Core 3.1.6 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.31603), X64 RyuJIT


|                    Method |     Mean |     Error |    StdDev |   Median |       Q3 |      P80 |      P85 |      P90 |      P95 |     P100 |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|----------:|-------:|------:|------:|----------:|
| CreateSignatureGeneration | 4.627 us | 0.2003 us | 0.5683 us | 4.338 us | 4.835 us | 5.073 us | 5.226 us | 5.402 us | 6.054 us | 6.398 us | 216,126.1 | 0.3815 |     - |     - |   1.57 KB |
|   ReadSignatureGeneration | 4.428 us | 0.0907 us | 0.2276 us | 4.361 us | 4.516 us | 4.559 us | 4.652 us | 4.811 us | 4.956 us | 5.058 us | 225,846.0 | 0.3891 |     - |     - |    1.6 KB |

|                    Method |     Mean |     Error |    StdDev |   Median |       Q3 |      P80 |      P85 |      P90 |      P95 |     P100 |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|----------:|-------:|------:|------:|----------:|
| CreateSignatureGeneration | 4.967 us | 0.0997 us | 0.2610 us | 4.884 us | 5.122 us | 5.157 us | 5.240 us | 5.336 us | 5.488 us | 5.765 us | 201,334.0 | 0.3433 |     - |     - |   1.41 KB |
|   ReadSignatureGeneration | 4.847 us | 0.0940 us | 0.1045 us | 4.851 us | 4.927 us | 4.930 us | 4.940 us | 4.962 us | 5.004 us | 5.036 us | 206,299.5 | 0.3433 |     - |     - |   1.43 KB |


BenchmarkDotNet=v0.12.0, OS=ubuntu 18.04
Intel Xeon Platinum 8171M CPU 2.60GHz, 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.301
  [Host]     : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT

|                    Method |     Mean |     Error |    StdDev |       Q3 |      P80 |      P85 |      P90 |      P95 |     P100 |      Op/s | Completed Work Items | Lock Contentions |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|---------:|----------:|---------------------:|-----------------:|-------:|------:|------:|----------:|
| CreateSignatureGeneration | 5.302 us | 0.0665 us | 0.0555 us | 5.344 us | 5.345 us | 5.353 us | 5.371 us | 5.396 us | 5.425 us | 188,590.9 |               0.0000 |                - | 0.0763 |     - |     - |   1.48 KB |
|   ReadSignatureGeneration | 5.369 us | 0.0743 us | 0.0695 us | 5.412 us | 5.420 us | 5.445 us | 5.462 us | 5.479 us | 5.499 us | 186,251.0 |               0.0000 |                - | 0.0763 |     - |     - |   1.48 KB |


|                    Method |     Mean |     Error |    StdDev |       Q3 |      P80 |      P85 |      P90 |      P95 |     P100 |      Op/s | Completed Work Items | Lock Contentions |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|---------:|----------:|---------------------:|-----------------:|-------:|------:|------:|----------:|
| CreateSignatureGeneration | 6.078 us | 0.0925 us | 0.0772 us | 6.145 us | 6.148 us | 6.159 us | 6.165 us | 6.169 us | 6.172 us | 164,537.6 |               0.0000 |                - | 0.0687 |     - |     - |    1.3 KB |
|   ReadSignatureGeneration | 6.317 us | 0.0777 us | 0.0726 us | 6.371 us | 6.375 us | 6.388 us | 6.396 us | 6.420 us | 6.465 us | 158,313.6 |               0.0000 |                - | 0.0687 |     - |     - |   1.33 KB |