@kirankumarkolli 2020-07-22
### AUTH EXCLUDED

|              Method |     Median |        P90 |         P95 | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |-----------:|-----------:|------------:|------:|------:|------:|----------:|
|          CreateItem |   993.2 us | 1,301.7 us |  1,376.3 us |     - |     - |     - |   38.7 KB |
|    DeleteItemExists |   800.5 us | 1,081.3 us |  1,126.7 us |     - |     - |     - |  38.71 KB |
| DeleteItemNotExists | 1,422.2 us | 2,265.5 us |  2,697.1 us |     - |     - |     - |   55.7 KB |
|            ReadFeed | 1,006.0 us | 1,725.8 us |  2,062.9 us |     - |     - |     - |  52.02 KB |
|      ReadItemExists | 1,491.2 us | 3,168.5 us |  3,419.7 us |     - |     - |     - |  43.68 KB |
|   ReadItemNotExists | 1,668.4 us | 3,304.9 us |  3,599.1 us |     - |     - |     - |  60.65 KB |
|          UpdateItem | 1,473.7 us | 2,930.9 us |  3,454.5 us |     - |     - |     - |  38.93 KB |
|          UpsertItem | 1,395.8 us | 2,971.6 us |  3,205.8 us |     - |     - |     - |  38.84 KB |

### AUTH BEFORE PERF CHANGES
|              Method |     Median |        P90 |        P95 | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |-----------:|-----------:|-----------:|------:|------:|------:|----------:|
|          CreateItem |   966.3 us | 1,559.0 us | 1,696.5 us |     - |     - |     - |  40.57 KB |
|    DeleteItemExists |   828.2 us | 1,233.3 us | 1,369.3 us |     - |     - |     - |  40.67 KB |
| DeleteItemNotExists |   924.1 us | 1,434.4 us | 1,563.2 us |     - |     - |     - |  57.62 KB |
|            ReadFeed | 1,013.1 us | 1,598.4 us | 1,852.3 us |     - |     - |     - |  53.64 KB |
|      ReadItemExists |   953.9 us | 2,114.5 us | 2,307.2 us |     - |     - |     - |  45.61 KB |
|   ReadItemNotExists | 1,015.2 us | 1,794.7 us | 2,146.4 us |     - |     - |     - |  62.56 KB |
|          UpdateItem | 1,005.8 us | 1,598.0 us | 1,846.0 us |     - |     - |     - |  40.88 KB |
|          UpsertItem |   829.0 us | 1,656.9 us | 1,905.7 us |     - |     - |     - |  40.73 KB |

### AUTH AFTER PERF CHANGES
|              Method |     Median |        P90 |        P95 | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |-----------:|-----------:|-----------:|------:|------:|------:|----------:|
|          CreateItem |   879.2 us | 1,583.2 us | 1,792.8 us |     - |     - |     - |  45.02 KB |
|    DeleteItemExists |   832.0 us | 1,389.4 us | 1,689.5 us |     - |     - |     - |   44.5 KB |
| DeleteItemNotExists | 1,567.2 us | 1,976.8 us | 2,191.8 us |     - |     - |     - |  61.48 KB |
|            ReadFeed | 1,416.6 us | 3,028.2 us | 3,230.3 us |     - |     - |     - |  57.52 KB |
|      ReadItemExists | 1,501.0 us | 3,445.5 us | 3,695.7 us |     - |     - |     - |  49.45 KB |
|   ReadItemNotExists |   993.8 us | 1,595.4 us | 1,697.7 us |     - |     - |     - |  66.45 KB |
|          UpdateItem | 1,164.2 us | 3,025.6 us | 3,383.5 us |     - |     - |     - |  44.69 KB |
|          UpsertItem | 1,007.4 us | 1,721.6 us | 2,084.4 us |     - |     - |     - |  44.63 KB |