@kirankumarkolli 2020-07-22
NOTE: Runs are from dev machine, latenciess might be unreliable. Lets just focus on allocations and GC's

Surprisingly allocations are more higher with the new implementation.

### Host context
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.302
  [Host]  : .NET Core 3.0.3 (CoreCLR 4.700.20.6603, CoreFX 4.700.20.6701), X64 RyuJIT