@kirankumarkolli 2020-07-22
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.302
  [Host]  : .NET Core 3.0.3 (CoreCLR 4.700.20.6603, CoreFX 4.700.20.6701), X64 RyuJIT
  LongRun : .NET Core 3.0.3 (CoreCLR 4.700.20.6603, CoreFX 4.700.20.6701), X64 RyuJIT

Job=LongRun  IterationCount=100  LaunchCount=3
WarmupCount=15

### AUTH BEFORE PERF CHANGES
|              Method |   Median |        P90 |        P95 |       P100 | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|-----------:|-----------:|-----------:|------:|------:|------:|----------:|
|          CreateItem | 839.9 us | 1,299.0 us | 1,579.2 us | 2,028.6 us |     - |     - |     - |  40.55 KB |
|    DeleteItemExists | 837.2 us | 1,579.1 us | 1,780.3 us | 2,267.2 us |     - |     - |     - |  40.63 KB |
| DeleteItemNotExists | 860.9 us | 1,248.8 us | 1,347.8 us | 2,112.5 us |     - |     - |     - |  57.58 KB |
|            ReadFeed | 898.3 us | 1,805.2 us | 2,096.3 us | 2,539.3 us |     - |     - |     - |  53.62 KB |
|      ReadItemExists | 924.4 us | 1,813.8 us | 2,104.7 us | 2,803.6 us |     - |     - |     - |  45.63 KB |
|   ReadItemNotExists | 917.1 us | 1,359.9 us | 1,470.6 us | 1,954.0 us |     - |     - |     - |  62.56 KB |
|          UpdateItem | 762.6 us | 1,183.4 us | 1,532.9 us | 1,958.7 us |     - |     - |     - |  40.87 KB |
|          UpsertItem | 806.9 us | 1,184.7 us | 1,286.7 us | 1,813.2 us |     - |     - |     - |  40.71 KB |

### AUTH AFTER PERF CHANGES
|              Method |     Median |        P90 |        P95 |       P100 | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |-----------:|-----------:|-----------:|-----------:|------:|------:|------:|----------:|
|          CreateItem | 1,305.8 us | 2,676.5 us | 3,137.0 us | 4,100.0 us |     - |     - |     - |  44.29 KB |
|    DeleteItemExists | 1,404.5 us | 2,970.9 us | 3,461.6 us | 4,130.9 us |     - |     - |     - |   44.5 KB |
| DeleteItemNotExists |   968.8 us | 1,857.0 us | 2,157.2 us | 2,761.7 us |     - |     - |     - |  61.46 KB |
|            ReadFeed |   962.0 us | 1,838.8 us | 2,047.7 us | 2,638.8 us |     - |     - |     - |  57.48 KB |
|      ReadItemExists |   829.8 us | 1,862.6 us | 2,119.4 us | 2,532.4 us |     - |     - |     - |  49.45 KB |
|   ReadItemNotExists |   914.2 us | 1,428.9 us | 1,768.4 us | 2,422.3 us |     - |     - |     - |  66.41 KB |
|          UpdateItem |   793.4 us | 1,332.0 us | 1,565.9 us | 2,167.4 us |     - |     - |     - |   44.7 KB |
|          UpsertItem |   759.5 us | 1,243.7 us | 1,401.4 us | 1,745.7 us |     - |     - |     - |  44.55 KB |