@visridha 2020-07-23
Latest run:
|                    Method |     Mean |     Error |    StdDev |       Q3 |      P80 |      P85 |      P90 |      P95 |     P100 |      Op/s | Completed Work Items | Lock Contentions |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|---------:|----------:|---------------------:|-----------------:|-------:|------:|------:|----------:|
| CreateSignatureGeneration | 3.219 us | 0.0883 us | 0.2604 us | 3.376 us | 3.428 us | 3.461 us | 3.585 us | 3.694 us | 3.858 us | 310,645.4 |               0.0000 |                - | 0.0534 |     - |     - |     896 B |
|   ReadSignatureGeneration | 3.192 us | 0.0820 us | 0.2419 us | 3.392 us | 3.427 us | 3.452 us | 3.540 us | 3.582 us | 3.797 us | 313,263.5 |               0.0000 |                - | 0.0534 |     - |     - |     896 B |